Move Rhs2116Stimulus and Rhs2116StimulusSequence to their own files
